use word renders instead of images
